LocalServerStress: Cleanup and consolidation
