+++ packages/test/local-server-stress-tests/src/test/localServerStress.spec.ts
@@ -32,16 +32,16 @@ describe("Local Server Stress", () => {
 	};
 
 	createLocalServerStressSuite(model, {
-		defaultTestCount: 100,
-		// skipMinimization: true,
-		// Uncomment to replay a particular seed.
-		// replay: 93,
-		// only: [30, 69],
+		defaultTestCount: 200,
 		saveFailures,
-		// saveSuccesses,
-		configurations: { "Fluid.Container.enableOfflineLoad": true },
+		configurations: {
+			"Fluid.Container.enableOfflineLoad": true,
+			"Fluid.ContainerRuntime.EnableRollback": true,
+		},
+		// skipMinimization: true,
+		// Use skip, replay, and only properties to control which seeds run.
 		skip: [
-			...[62], // shared array error
+			124, // directory 0xc38
 		],
 	});
 });
